backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Ilya Moiseev <ilya.moiseev95@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
asgiref = "3.2.10"
requests = "2.26.0"
Django = "2.2.16"
django-filter = "2.4.0"
djangorestframework = "3.12.4"
djangorestframework-simplejwt = "4.8.0"
gunicorn = "20.0.4"
psycopg2-binary = "2.8.6"
djoser = "2.1.0"
PyJWT = "2.1.0"
pytz = "2020.1"
sqlparse = "0.3.1"
six = "1.15.0"
flake8 = "^4.0.1"
django-colorfield = "^0.6.3"
django-extra-fields = "^3.0.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.1.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
